ajustes

+++ src/main/java/io/github/mendjoy/gymJourneyAPI/utils/ValidationUtils.java
@@ -4,7 +4,7 @@ import io.github.mendjoy.gymJourneyAPI.exception.GymJourneyException;
 
 public class ValidationUtils {
 
-    public static void validateIdNotNull(Integer id, String entityName) {
+    public static void validateIdNotNull(Long id, String entityName) {
         if (id == null) {
             throw GymJourneyException.badRequest("O id do(a) " + entityName + " não pode ser nulo!");
         }
